Refactor and clean up model and data utility files
- Removed unused imports and consolidated import statements in `__init__.py` files across various model packages, enhancing code clarity and organization.
- Commented out logging statements in `ml_feature_loader.py` to reduce verbosity during execution while maintaining the ability to restore them if needed.
- Updated data type for categorical indices in `realmlp_predictor.py` from `long` to `bfloat16` for improved memory efficiency.
- Introduced early stopping parameters and enhanced logging in `realmlp_trainer.py` to facilitate better training monitoring and model checkpointing.
- Cleaned up the `realmlp_architecture.py` and `realmlp_preprocessing.py` files by removing unnecessary comments and improving code readability.

+++ src/models/time_series/realmlp/realmlp_architecture.py
@@ -109,7 +109,7 @@ class RealMLPModule(nn.Module):
             # Fallback: if categorical indices are not provided during inference/logging,
             # use a zero (OOV) index vector to preserve expected trunk input dimensionality
             if cat_idx is None:
-                cat_idx = torch.zeros(z.size(0), dtype=torch.long, device=z.device)
+                cat_idx = torch.zeros(z.size(0), dtype=torch.bfloat16, device=z.device)
             if cat_idx is not None:
                 # Ensure 1D indices
                 if cat_idx.dim() > 1:
